# twitter-clone/openapi.yaml
openapi: 3.0.3
info:
  title: Twitter Clone API
  description: A production-ready Twitter clone API with microservices architecture
  version: 1.0.0
  contact:
    name: API Support
    url: https://twitter-clone.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.twitter-clone.com
    description: Production server
  - url: https://staging-api.twitter-clone.com
    description: Staging server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/tweets/{tweetId}:
    get:
      tags: [Tweets]
      summary: Get a tweet
      operationId: getTweet
      parameters:
        - name: tweetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tweet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TweetDetails'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tweets]
      summary: Delete a tweet
      operationId: deleteTweet
      parameters:
        - name: tweetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tweet deleted
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        bio:
          type: string
          nullable: true
        avatar_url:
          type: string
          format: uri
          nullable: true
        follower_count:
          type: integer
          minimum: 0
        following_count:
          type: integer
          minimum: 0
        tweet_count:
          type: integer
          minimum: 0
        verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Tweet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        content:
          type: string
        media_ids:
          type: array
          items:
            type: string
            format: uuid
        reply_to_id:
          type: string
          format: uuid
          nullable: true
        retweet_of_id:
          type: string
          format: uuid
          nullable: true
        like_count:
          type: integer
          minimum: 0
        retweet_count:
          type: integer
          minimum: 0
        reply_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
    TweetDetails:
      allOf:
        - $ref: '#/components/schemas/Tweet'
        - type: object
          properties:
            author:
              $ref: '#/components/schemas/User'
            media_urls:
              type: array
              items:
                type: string
                format: uri
            is_liked:
              type: boolean
            is_retweeted:
              type: boolean
            is_bookmarked:
              type: boolean

    # Timeline Schemas
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...